Re-using connections for benchmarking tests
